requirements.txt
requests==2.32.5
pandas==2.3.3
sqlalchemy==2.0.45
supabase==2.25.0
python-dotenv==1.2.1